Updating cogs

+++ utils/helpers.py
@@ -23,13 +23,14 @@ def checkdigitarguments(args, default):
     return default
 
 
-async def message_handler(message, ctx, time=30, embed=False, delete=True):
+async def message_handler(message, ctx, time=30, embed=False, delete=True, deleteresponse=True):
     m = await ctx.send(embed=message) if embed else await ctx.send(message)
     if delete:
         await sleep(0)
         await ctx.message.delete()
-    await sleep(time)
-    await m.delete()
+    if deleteresponse:
+        await sleep(time)
+        await m.delete()
 
 
 async def message_denied(message, ctx, pm=False):
